New 3.0 GA Version of SAP Creation script
This is a GUI based Pyhon Script for Creation a SAP. Based on Support working by sophos network security group.

Basics:
But first the basics, what you need and what are the scripts good for.
We need Python 3.x Download for Windows.

After installation we can working with scripts or develop scripts.

for additional funktions e.g. Word or PDF export open Powershell as admin and put follow commands:
pip install python-docx
pip install reportlab

This is version 3.0 of the SAP Creation script. The new script, like the old one, is based on Python. However, now we have a GUI that allows us to enter data via copy and past into multi-line text fields.
There is export to word and as pdf. And we have an import function for old SAPs text files.

system requirements:

    Python 3.10 and higher

    for export we need additional phyton-docx and reportlab

 
new features:

Gui based

Multiline text fields

Export of SAP to Microsoft Word*

Export of SAP to PDF*

Dropdown fields

New fields

Deleting the fields after creating the SAP so you can create the next sap without closing the program.

Import of old SAP files (txt) created with this version. So you can change things, add and create an update.

Empty fields are ignored and omitted when creating. * This only applies to the TEXT file; when exporting to Word or PDF, all fields are also exported to the blank.